api/requirements.txt
#python 3.8
asgiref==3.3.4
certifi==2020.12.5
chardet==3.0.4
dateutils==0.6.7
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
idna==2.8
Markdown==3.3.4
oauthlib==3.1.0
psycopg2==2.9.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.22.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
uWSGI==2.0.19.1

